# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: RALD ELIMU — African Education Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schools:
    get:
      operationId: listSchools
      tags: [schools]
      summary: List all schools
      parameters:
        - in: query
          name: country
          schema:
            type: string
        - in: query
          name: type
          schema:
            $ref: "#/components/schemas/SchoolType"
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of schools
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolListResponse"
    post:
      operationId: createSchool
      tags: [schools]
      summary: Register a new school
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSchoolBody"
      responses:
        "201":
          description: Created school
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /schools/{schoolId}:
    get:
      operationId: getSchool
      tags: [schools]
      summary: Get a school by ID
      parameters:
        - $ref: "#/components/parameters/schoolId"
      responses:
        "200":
          description: School
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateSchool
      tags: [schools]
      summary: Update school details
      parameters:
        - $ref: "#/components/parameters/schoolId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSchoolBody"
      responses:
        "200":
          description: Updated school
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Students ───────────────────────────────────────────────────────────────
...
  /schools/{schoolId}/parents:
    get:
      operationId: listParents
      tags: [parents]
      summary: List parents in a school
      parameters:
        - $ref: "#/components/parameters/schoolId"
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of parents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParentListResponse"
    post:
      operationId: addParent
      tags: [parents]
      summary: Add a parent/guardian
      parameters:
        - $ref: "#/components/parameters/schoolId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddParentBody"
      responses:
        "201":
          description: Added parent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parent"
  /parents/{parentId}:
    get:
      operationId: getParent
      tags: [parents]
      summary: Get a parent profile
      parameters:
        - $ref: "#/components/parameters/parentId"
      responses:
        "200":
          description: Parent profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parent"
  /parents/{parentId}/children:
    get:
      operationId: getParentChildren
      tags: [parents]
      summary: Get all children of a parent
      parameters:
        - $ref: "#/components/parameters/parentId"
      responses:
        "200":
          description: Children list
          content:
            application/json:
              schema:
                type: object
                properties:
                  children:
                    type: array
                    items:
                      $ref: "#/components/schemas/Student"
                required: [children]

  # ── Classes ────────────────────────────────────────────────────────────────
...
  /schools/{schoolId}/fees:
    get:
      operationId: listFeeStructures
      tags: [fees]
      summary: List fee structures for a school
      parameters:
        - $ref: "#/components/parameters/schoolId"
        - in: query
          name: academicYear
          schema:
            type: string
        - in: query
          name: academicTerm
          schema:
            type: string
      responses:
        "200":
          description: Fee structures
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeStructureListResponse"
    post:
      operationId: createFeeStructure
      tags: [fees]
      summary: Create a fee structure
      parameters:
        - $ref: "#/components/parameters/schoolId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeeStructureBody"
      responses:
        "201":
          description: Created fee structure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeStructure"
  /fees/{feeId}:
    get:
      operationId: getFeeStructure
      tags: [fees]
      summary: Get a fee structure
      parameters:
        - $ref: "#/components/parameters/feeId"
      responses:
        "200":
          description: Fee structure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeStructure"
  /fees/{feeId}/payments:
    get:
      operationId: listFeePayments
      tags: [fees]
      summary: List payments for a fee structure
      parameters:
        - $ref: "#/components/parameters/feeId"
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/PaymentStatus"
      responses:
        "200":
          description: Fee payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentListResponse"

  # ── Payments ───────────────────────────────────────────────────────────────
  /payments:
    post:
      operationId: recordPayment
      tags: [payments]
      summary: Record a fee payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordPaymentBody"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /students/{studentId}/payments:
    get:
      operationId: getStudentPayments
      tags: [payments]
      summary: Get all payments for a student
      parameters:
        - $ref: "#/components/parameters/studentId"
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/PaymentStatus"
        - in: query
          name: academicYear
          schema:
            type: string
      responses:
        "200":
          description: Student payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentListResponse"
  /payments/{paymentId}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get a specific payment
      parameters:
        - in: path
          name: paymentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ── Announcements ──────────────────────────────────────────────────────────
...
components:
  parameters:
    schoolId:
      in: path
      name: schoolId
      required: true
      schema:
        type: string
    studentId:
      in: path
      name: studentId
      required: true
      schema:
        type: string
...
    parentId:
      in: path
      name: parentId
      required: true
      schema:
        type: string
...
    feeId:
      in: path
      name: feeId
      required: true
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [error]
...
    SchoolType:
      type: string
      enum: [kindergarten, primary, secondary, university, polytechnic, vocational]
    StudentStatus:
      type: string
      enum: [active, graduated, transferred, suspended, withdrawn]
...
    AcademicLevel:
      type: string
      enum: [kindergarten, primary, secondary, university]
...
    PaymentStatus:
      type: string
      enum: [pending, completed, failed, refunded, partial]
    PaymentMethod:
      type: string
      enum: [cash, bank_transfer, card, mobile_money, scholarship, grant]
...
    School:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          $ref: "#/components/schemas/SchoolType"
        country:
          type: string
        state:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        accreditationStatus:
          type: string
          nullable: true
        verificationStatus:
          type: string
          enum: [unverified, pending, verified]
        enrollmentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, type, country, verificationStatus, enrollmentCount, createdAt, updatedAt]
    SchoolListResponse:
      type: object
      properties:
        schools:
          type: array
          items:
            $ref: "#/components/schemas/School"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [schools, total, page, limit]
    CreateSchoolBody:
      type: object
      properties:
        name:
          type: string
        type:
          $ref: "#/components/schemas/SchoolType"
        country:
          type: string
        state:
          type: string
        city:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        logoUrl:
          type: string
      required: [name, type, country]
    UpdateSchoolBody:
      type: object
      properties:
        name:
          type: string
        type:
          $ref: "#/components/schemas/SchoolType"
        country:
          type: string
        state:
          type: string
        city:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        logoUrl:
          type: string
        accreditationStatus:
          type: string

    # ── Student ─────────────────────────────────────────────────────────────
    Student:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        studentNumber:
          type: string
        schoolId:
          type: string
        classId:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          format: date
          nullable: true
        gender:
          type: string
          nullable: true
        enrollmentDate:
          type: string
          format: date
        academicLevel:
          $ref: "#/components/schemas/AcademicLevel"
        status:
          $ref: "#/components/schemas/StudentStatus"
        profileImageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, firstName, lastName, studentNumber, schoolId, enrollmentDate, academicLevel, status, createdAt, updatedAt]
...
    Parent:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        schoolId:
          type: string
        occupation:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        childrenIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, firstName, lastName, schoolId, childrenIds, createdAt, updatedAt]
    ParentListResponse:
      type: object
      properties:
        parents:
          type: array
          items:
            $ref: "#/components/schemas/Parent"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [parents, total, page, limit]
    AddParentBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        occupation:
          type: string
        address:
          type: string
        studentIds:
          type: array
          items:
            type: string
      required: [firstName, lastName]

    # ── Class ───────────────────────────────────────────────────────────────
...
    FeeStructure:
      type: object
      properties:
        id:
          type: string
        schoolId:
          type: string
        name:
          type: string
        amount:
          type: number
        currency:
          type: string
        academicYear:
          type: string
        academicTerm:
          type: string
          nullable: true
        academicLevel:
          $ref: "#/components/schemas/AcademicLevel"
        dueDate:
          type: string
          format: date
          nullable: true
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, schoolId, name, amount, currency, academicYear, academicLevel, createdAt, updatedAt]
    FeeStructureListResponse:
      type: object
      properties:
        fees:
          type: array
          items:
            $ref: "#/components/schemas/FeeStructure"
        total:
          type: integer
      required: [fees, total]
    CreateFeeStructureBody:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
        currency:
          type: string
          default: NGN
        academicYear:
          type: string
        academicTerm:
          type: string
        academicLevel:
          $ref: "#/components/schemas/AcademicLevel"
        dueDate:
          type: string
          format: date
        description:
          type: string
      required: [name, amount, academicYear, academicLevel]
    Payment:
      type: object
      properties:
        id:
          type: string
        feeStructureId:
          type: string
        studentId:
          type: string
        studentName:
          type: string
          nullable: true
        amountPaid:
          type: number
        currency:
          type: string
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        reference:
          type: string
        status:
          $ref: "#/components/schemas/PaymentStatus"
        paidAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, feeStructureId, studentId, amountPaid, currency, paymentMethod, reference, status, createdAt]
    PaymentListResponse:
      type: object
      properties:
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
        totalAmount:
          type: number
      required: [payments, total, totalAmount]
    RecordPaymentBody:
      type: object
      properties:
        feeStructureId:
          type: string
        studentId:
          type: string
        amountPaid:
          type: number
        currency:
          type: string
          default: NGN
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        reference:
          type: string
        status:
          $ref: "#/components/schemas/PaymentStatus"
        paidAt:
          type: string
          format: date-time
      required: [feeStructureId, studentId, amountPaid, paymentMethod]

    # ── Announcement ────────────────────────────────────────────────────────
...